refactor(db): align migrations and JPA to BIGINT identifiers
Standardize PKs and FKs as BIGSERIAL/BIGINT across Flyway scripts; fix V10 candidate_status seed (UUID vs integer). Split candidatos status assignment via V11/V14. Turn V7 into a no-op.

BaseEntity.id and related domain fields, ports, DTOs, and repositories use Long. SpringRefreshTokenRepository extends BaseRepository and implements revokeAllByUserId; LoginUseCase is transactional for revoke + save. RoleRepositoryAdapter delegates findById to the Spring repository.

BREAKING CHANGE: existing databases must be recreated (e.g. drop Docker volume) before running migrations.

Made-with: Cursor

+++ src/main/java/com/seletoai/adapters/persistence/analytics/AlertRepositoryAdapter.java
@@ -14,7 +14,7 @@ public class AlertRepositoryAdapter implements AlertRepositoryPort {
   private final SpringAlertRepository repository;
 
   @Override
-  public List<Alert> findByProcessIdAndResolvedIsFalse(Integer processId) {
+  public List<Alert> findByProcessIdAndResolvedIsFalse(Long processId) {
     return repository.findByProcessIdAndResolvedIsFalse(processId);
   }
 }
