+++ src/main/java/com/seletoai/adapters/persistence/userRole/SpringUserRoleRepository.java
@@ -10,5 +10,5 @@ public interface SpringUserRoleRepository
 
   List<UserRole> findByUserId(Long userId);
 
-  List<UserRole> findByRoleId(Integer roleId);
+  List<UserRole> findByRoleId(Long roleId);
 }
