+++ src/main/java/com/seletoai/core/domain/analytics/Alert.java
@@ -11,7 +11,7 @@ import lombok.*;
 public class Alert extends BaseEntity {
 
   @Column(name = "process_id", nullable = false)
-  private Integer processId;
+  private Long processId;
 
   private String type;
   private String message;
